Jdk version workaround (#89)
* EE-6630 Fix findbugs error

* Workaround for issue on jdk < 1.8.0._131

The compiler mistakenly thinks that the exception thrown by RetryTemplate.execute
should be caught on jdk < 1.8.0_131.  We are catching and rethrowing the
exception to workaround the issue.  This can be removed once we are confident
that all jdks used in builds are on a later version.

* Workaround for issue on jdk < 1.8.0._131

The compiler mistakenly thinks that the exception thrown in the lambda
should be caught on jdk < 1.8.0_131.  We are catching and rethrowing the
exception to workaround the issue.  This can be removed once we are confident
that all jdks used in builds are on a later version.