EE-6635 Use seperate rest template per service called
We need a rest template per service so that we can control the timeouts
for each service we call.

NOTE as we are no longer sharing a rest template, I removed the media type
HAL_JSON and the apache clientHttpRestFactory from the audit and access code
rest templates as they should not be required.  If they start breaking,
the first thing to try would be to put them back in.