EE-7402 Refactored the name matching permutations code in order to make it possible to have surnames with multiple words in (e.g. "Van Damme").
The HmrcHateoasClient used to depend on the names being a single string
of two names separated by a space. It would then split on this space in
order to determine the first and last names for the HMRC API call. I
have now changed it so that it instead expects a simple class called
PersonName which has 2 fields, firstname and surname. This will make
facilitate contacting the HMRC API with multi-word surnames although this
has yet to be implemented.